+++ src/ipmdchecker.ts
@@ -397,29 +397,6 @@ export class IpmdChecker {
             exifDataSet = true;
           }
         }
-        // special case for the IPTC property title: an Exif ImageDescription value without spaces makes one
-        /* IPTC Title--Exif ImageDescription mapping cancelled
-        if (etExifId === "IFD0:ImageDescription" && refPropId === "title") {
-          if (testImgEtPmd.hasOwnProperty(etExifId)) {
-            const testValue: string = testImgEtPmd[etExifId];
-            if (!testValue.includes(" ")) {
-              this._ipmdStateData.setFsData(
-                1,
-                refPropId +
-                  this._lsep +
-                  icc.ipmdcrSData +
-                  this._lsep +
-                  icc.ipmdcrSDexif
-              );
-              propVresult[icc.ipmdcrVexif] = testValue;
-              exifValue = testValue;
-              exifDataSet = true;
-            } else {
-              exifDataSet = true; // = in fact not setting a value
-            }
-          }
-        }
-        */
         // special case: Exif Tag Artist, make its value a single item array to compare properly
         if (etExifId === "IFD0:Artist") {
           if (testImgEtPmd.hasOwnProperty(etExifId)) {
@@ -480,10 +457,13 @@ export class IpmdChecker {
                 case "dateCreated":
                   let xmptestval: string = xmpValue;
                   if (xmptestval.includes("Z")) {
-                    // pre code: .substr(0, xmptestval.length - 1);
-                    xmptestval = xmptestval.substring(0, xmptestval.length);
+                    xmptestval =
+                      xmptestval.substring(0, xmptestval.length) + "+00:00";
                   }
                   xmpIimAreEqual = xmptestval === iimValue;
+                  break;
+                case "intellectualGenre":
+                  xmpIimAreEqual = "000:" + xmpValue === iimValue;
                   break;
                 default:
                   xmpIimAreEqual = xmpValue === iimValue;
@@ -763,7 +743,7 @@ export class IpmdChecker {
       testDataValue,
       false
     );
-    const compareResultRows: CompareResultRow[] = []; // this will be retured as result of this method
+    const compareResultRows: CompareResultRow[] = []; // this will be returned as result of this method
     // interate across the top level properties in the values of the reference file
     for (let idx = 0; idx < refValueIpmdIds.length; idx++) {
       const refIpmdId: string = refValueIpmdIds[idx];