+++ src/ipmdsetter.ts
@@ -2,7 +2,6 @@ import fs from "fs";
 import * as icc from "./constants";
 import { MdStruct } from "./incommon";
 import * as util1 from "./utilities1";
-import { ExifParts } from "./utilities1";
 import FixedStructureData from "./fixed_structure_data";
 import * as valmap from "./valuemapper";
 
@@ -106,7 +105,7 @@ export class IpmdSetter {
         propoccurrence === icc.itgPropoccurSingle &&
         typeof ipmdPropValue === "string"
       ) {
-        etPropValue = util1.xmpIsoDatetime2etDatetime(ipmdPropValue);
+        etPropValue = util1.xmpIsoDatetime2etDatetime(ipmdPropValue); // data type: EtDateTimeVariants
       }
       // datatype "number"
       if (
@@ -191,7 +190,18 @@ export class IpmdSetter {
                   altstring.substring(langEndIdx + 1);
               }
             });
-          } else etJson[refPropData[icc.itgEtXmp]] = etPropValue;
+          } else {
+            switch (refPropId) {
+              case "dateCreated":
+                if (etPropValue["xmpDateTime"] != null)
+                  etJson[refPropData[icc.itgEtXmp]] =
+                    etPropValue["xmpDateTime"];
+                break;
+              default:
+                etJson[refPropData[icc.itgEtXmp]] = etPropValue;
+                break;
+            }
+          }
         }
       }
       // set IIM value(s)
@@ -199,16 +209,11 @@ export class IpmdSetter {
         if (etPropValue !== undefined) {
           switch (refPropId) {
             case "dateCreated":
-              const exifParts: ExifParts =
-                util1.etDatetime2ExifParts(etPropValue);
-              if (exifParts.dateTime !== null) {
-                const etPropValues: string[] = exifParts.dateTime.split(" ");
-                etJson["IPTC:DateCreated"] = etPropValues[0];
-                if (etPropValues.length > 1) {
-                  etJson["IPTC:TimeCreated"] = etPropValues[1];
+              if (etPropValue["iimDate"] !== null) {
+                etJson["IPTC:DateCreated"] = etPropValue["iimDate"];
+                if (etPropValue["iimTime"] !== null) {
+                  etJson["IPTC:TimeCreated"] = etPropValue["iimTime"];
                 }
-                if (exifParts.tzOffset !== null)
-                  etJson["IPTC:TimeCreated"] += exifParts.tzOffset;
               }
               break;
             case "subjectCodes":
@@ -228,14 +233,16 @@ export class IpmdSetter {
         if (etPropValue !== undefined) {
           switch (refPropId) {
             case "dateCreated":
-              const exifParts: ExifParts =
-                util1.etDatetime2ExifParts(etPropValue);
-              if (exifParts.dateTime !== null)
-                etJson["ExifIFD:DateTimeOriginal"] = exifParts.dateTime;
-              if (exifParts.subSeconds !== null)
-                etJson["ExifIFD:SubSecTimeOriginal"] = exifParts.subSeconds;
-              if (exifParts.tzOffset !== null)
-                etJson["ExifIFD:OffsetTimeOriginal"] = exifParts.tzOffset;
+              if (etPropValue["exifDateTime"] !== null) {
+                etJson["ExifIFD:DateTimeOriginal"] =
+                  etPropValue["exifDateTime"];
+                if (etPropValue["exifSubSeconds"] !== null)
+                  etJson["ExifIFD:SubSecTimeOriginal"] =
+                    etPropValue["exifSubSeconds"];
+                if (etPropValue["exifTzOffset"] !== null)
+                  etJson["ExifIFD:OffsetTimeOriginal"] =
+                    etPropValue["exifTzOffset"];
+              }
               break;
             case "imageRegion": // the mapping to Exif's Subject Area is virtual
               break;
@@ -411,7 +418,19 @@ export class IpmdSetter {
                     altstring.substring(langEndIdx + 1);
                 }
               });
-            } else etJsonStruct[refPropData[icc.itgEtTag]] = etPropValue;
+            } else {
+              switch (refPropId) {
+                case "dateCreated":
+                  if (etPropValue["xmpDateTime"] != null)
+                    etJsonStruct[refPropData[icc.itgEtTag]] =
+                      etPropValue["xmpDateTime"];
+                  break;
+                default:
+                  etJsonStruct[refPropData[icc.itgEtTag]] = etPropValue;
+                  break;
+              }
+            }
+            // etJsonStruct[refPropData[icc.itgEtTag]] = etPropValue;
           }
         }
       }